+++ frontend/components/page/pwa.php
@@ -1,4 +1,4 @@
-<link rel="manifest" href="<?= get_public_url('manifest.json') ?>">
+<link rel="manifest" href="<?= get_public_url(MANIFEST_FILE) ?>">
 
 <meta name="theme-color" media="(prefers-color-scheme: light)" content="white">
 <meta name="theme-color" media="(prefers-color-scheme: dark)"  content="black">